+++ app/tests/Unit/Bakery/AssetsViteCommandTest.php
@@ -18,7 +18,7 @@ use phpmock\MockBuilder;
 use phpmock\mockery\PHPMockery;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
-use UserFrosting\Sprinkle\Core\Bakery\AssetsWebpackCommand;
+use UserFrosting\Sprinkle\Core\Bakery\AssetsViteCommand;
 use UserFrosting\Sprinkle\Core\Bakery\Helper\ShellCommandHelper;
 use UserFrosting\Sprinkle\Core\Exceptions\VersionCompareException;
 use UserFrosting\Sprinkle\Core\Validators\NodeVersionValidator;
@@ -27,18 +27,18 @@ use UserFrosting\Testing\BakeryTester;
 use UserFrosting\Testing\ContainerStub;
 
 /**
- * Test `assets:webpack` command.
+ * Test `assets:vite` command.
  *
  * Warning : This test doesn't fully test the output format.
  */
-class AssetsWebpackCommandTest extends TestCase
+class AssetsViteCommandTest extends TestCase
 {
     use MockeryPHPUnitIntegration;
 
     public function testCommand(): void
     {
         // Mock built-in function from main class
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
         PHPMockery::mock($namespace, 'file_exists')->andReturn(true, false);
@@ -62,20 +62,20 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
         $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('npm run webpack:dev', $result->getDisplay());
-        $this->assertStringContainsString('Webpack Encore run completed', $result->getDisplay());
+        $this->assertStringContainsString('npm run vite:dev', $result->getDisplay());
+        $this->assertStringContainsString('Vite command completed', $result->getDisplay());
     }
 
     public function testCommandProductionEnv(): void
     {
         // Mock built-in function from main class
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
         PHPMockery::mock($namespace, 'file_exists')->andReturn(true, false);
@@ -99,20 +99,20 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', 'production'); // Set production mode
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
         $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('npm run webpack:build', $result->getDisplay());
-        $this->assertStringContainsString('Webpack Encore run completed', $result->getDisplay());
+        $this->assertStringContainsString('npm run vite:build', $result->getDisplay());
+        $this->assertStringContainsString('Vite command completed', $result->getDisplay());
     }
 
     public function testCommandProduction(): void
     {
         // Mock built-in function from main class
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
         PHPMockery::mock($namespace, 'file_exists')->andReturn(true, false);
@@ -136,95 +136,20 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command, input: ['--production' => true]);
 
         // Assert some output
         $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('npm run webpack:build', $result->getDisplay());
-        $this->assertStringContainsString('Webpack Encore run completed', $result->getDisplay());
-    }
-
-    public function testCommandWatch(): void
-    {
-        // Mock built-in function from main class
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
-        $namespace = $reflection_class->getNamespaceName();
-        PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
-        PHPMockery::mock($namespace, 'file_exists')->andReturn(true, false);
-
-        // Mock passthru, from ShellCommandHelper
-        $reflection_class = new ReflectionClass(ShellCommandHelper::class);
-        $namespace = $reflection_class->getNamespaceName();
-        PHPMockery::mock($namespace, 'passthru')->andReturn(null);
-
-        // Set Validator mock
-        $node = Mockery::mock(NodeVersionValidator::class)
-            ->shouldReceive('validate')->andReturn(true)
-            ->getMock();
-        $npm = Mockery::mock(NpmVersionValidator::class)
-            ->shouldReceive('validate')->andReturn(true)
-            ->getMock();
-
-        // Set mock in CI and run command
-        $ci = ContainerStub::create();
-        $ci->set(NodeVersionValidator::class, $node);
-        $ci->set(NpmVersionValidator::class, $npm);
-        $ci->set('UF_MODE', '');
-
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
-        $result = BakeryTester::runCommand($command, input: ['--watch' => true]);
-
-        // Assert some output
-        $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('npm run webpack:watch', $result->getDisplay());
-        $this->assertStringContainsString('Webpack Encore run completed', $result->getDisplay());
-    }
-
-    public function testCommandWatchAndProduction(): void
-    {
-        // Mock built-in function from main class
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
-        $namespace = $reflection_class->getNamespaceName();
-        PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
-        PHPMockery::mock($namespace, 'file_exists')->andReturn(true, false);
-
-        // Mock passthru, from ShellCommandHelper
-        $reflection_class = new ReflectionClass(ShellCommandHelper::class);
-        $namespace = $reflection_class->getNamespaceName();
-        PHPMockery::mock($namespace, 'passthru')->andReturn(null);
-
-        // Set Validator mock
-        $node = Mockery::mock(NodeVersionValidator::class)
-            ->shouldReceive('validate')->andReturn(true)
-            ->getMock();
-        $npm = Mockery::mock(NpmVersionValidator::class)
-            ->shouldReceive('validate')->andReturn(true)
-            ->getMock();
-
-        // Set mock in CI and run command
-        $ci = ContainerStub::create();
-        $ci->set(NodeVersionValidator::class, $node);
-        $ci->set(NpmVersionValidator::class, $npm);
-        $ci->set('UF_MODE', '');
-
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
-        $result = BakeryTester::runCommand($command, input: ['--watch' => true, '--production' => true]);
-
-        // Assert some output
-        // N.B.: When both production & watch are used, production has priority
-        $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('npm run webpack:build', $result->getDisplay());
-        $this->assertStringContainsString('Webpack Encore run completed', $result->getDisplay());
+        $this->assertStringContainsString('npm run vite:build', $result->getDisplay());
+        $this->assertStringContainsString('Vite command completed', $result->getDisplay());
     }
 
     public function testCommandWithMissingFiles(): void
     {
         // Mock built-in error_get_last
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn('./foo');
         PHPMockery::mock($namespace, 'file_exists')->andReturn(false);
@@ -243,19 +168,19 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
         $this->assertSame(0, $result->getStatusCode());
-        $this->assertStringContainsString('./foo/webpack.config.js not found. Skipping.', $result->getDisplay());
+        $this->assertStringContainsString('Vite config not found. Skipping.', $result->getDisplay());
     }
 
     public function testCommandWithErrorInGetcwd(): void
     {
         // Mock built-in error_get_last
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn(false);
 
@@ -273,8 +198,8 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
@@ -296,8 +221,8 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
@@ -320,8 +245,8 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
@@ -331,7 +256,7 @@ class AssetsWebpackCommandTest extends TestCase
     public function testCommandWithNpmPassthruError(): void
     {
         // Mock built-in error_get_last
-        $reflection_class = new ReflectionClass(AssetsWebpackCommand::class);
+        $reflection_class = new ReflectionClass(AssetsViteCommand::class);
         $namespace = $reflection_class->getNamespaceName();
         PHPMockery::mock($namespace, 'getcwd')->andReturn('foo');
         PHPMockery::mock($namespace, 'file_exists')->andReturn(true);
@@ -366,13 +291,13 @@ class AssetsWebpackCommandTest extends TestCase
         $ci->set(NpmVersionValidator::class, $npm);
         $ci->set('UF_MODE', '');
 
-        /** @var AssetsWebpackCommand */
-        $command = $ci->get(AssetsWebpackCommand::class);
+        /** @var AssetsViteCommand */
+        $command = $ci->get(AssetsViteCommand::class);
         $result = BakeryTester::runCommand($command);
 
         // Assert some output
         $this->assertSame(1, $result->getStatusCode());
-        $this->assertStringContainsString('Webpack Encore run has failed', $result->getDisplay());
+        $this->assertStringContainsString('Vite command has failed', $result->getDisplay());
 
         // Disable mock manually
         $mock->disable();